+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # CasualOS Changelog
 
+## V1.0.27
+
+### Date: TBD
+
+### Changes:
+
+-   :rocket: Improvements
+
+    -   Added the `player.share(options)` function.
+        -   This will trigger the device's social share capabilities to share the given URL or text.
+        -   Note that this only works on Android and iOS phones and only works in response to some user action like a click.
+        -   `options` is an object with at least one of the following properties:
+            -   `url` - The URL to share. (optional)
+            -   `text` - The text to share. (optional)
+            -   `title` - The title of the document that is being shared. (optional)
+
 ## V1.0.26
 
 ### Date: 4/21/2020